Created a new component for FilterOption
