+++ src/Components/Notification.jsx
@@ -1,9 +1,12 @@
-import React from 'react'
+// src/components/Notification.jsx
+import React from 'react';
 
 const Notification = () => {
   return (
-    <div>Notification</div>
-  )
-}
+    <div className="bg-blue-100 border border-blue-400 text-blue-700 px-4 py-3 rounded relative" role="alert">
+      <span className="block sm:inline">This is a notification component.</span>
+    </div>
+  );
+};
 
-export default Notification
+export default Notification;
